remove distance to separate file
